Extract shared CSharp name formatter
Move C# identifier, namespace, arity, and simple type-name cleanup out of CompileBackSourceComposer into a shared metadata-side formatter used by the declaration writer and compile-back composer.
